add example and fix

+++ rio_viz/algorithm/index.py
@@ -20,8 +20,8 @@ class NormalizedIndex(BaseAlgorithm):
 
     def apply(self, img: ImageData) -> ImageData:
         """Normalized difference."""
-        b1 = img.data[0]
-        b2 = img.data[1]
+        b1 = img.data[0].astype("float32")
+        b2 = img.data[1].astype("float32")
 
         arr = numpy.where(img.mask, (b2 - b1) / (b2 + b1), 0)
